Updates to the auto_instrument example
- Renamed directory
- Fixed incorrect pragmas
- Added a few diagrams and screenshots
- Updated readme file
