--- auto-instrument/Makefile
@@ -1,6 +1,0 @@
-SRCS=main.cpp  
-LOCAL_CONFIG = -legup-config=config.tcl
-LEVEL = $(SHLS_ROOT_DIR)/examples
-
-include $(LEVEL)/Makefile.common
-HLS_INSTRUMENT_ENABLE=1
