+++ auto_instrument/assets/empty_signals_delay_0.png
