+++ auto_instrument/assets/empty_signals_delay_1.png
