+++ auto_instrument/assets/fifo_monitoring_delays_20_40_80_160.png
