+++ auto_instrument/assets/fifo_monitoring_delays_220_20_150_80.png
