+++ auto_instrument/assets/fifo_monitoring_executable_not_running.png
