+++ auto_instrument/assets/identify_gui.png
