+++ auto_instrument/assets/wave_template_grouping.png
