+++ auto_instrument/config.tcl
